+++ tests/cluster/local.go
@@ -56,7 +56,7 @@ func (ftry *LocalFactory) setup() error {
 	quotas := MakeRandomQuotas(ftry.params.NodeCount, ftry.params.StakeQuota)
 	genesis := &node.Genesis{
 		Validators: make([]string, len(keys)),
-		Quotas:     make([]float64, len(keys)),
+		Quotas:     make([]uint32, len(keys)),
 		WinSize:    ftry.params.WinSize,
 	}
 	for i, v := range keys {
